+++ src/TileMap.cpp
@@ -42,17 +42,17 @@ void TileMap::updateVertexArray() {
             }
 
             // Get texture position
-            unsigned int tu = m_tiles[i + j * width]->getTextureId() % (m_tileset->getSize().x / Constants::GRID_SIZE);
-            unsigned int tv = m_tiles[i + j * width]->getTextureId() / (m_tileset->getSize().x / Constants::GRID_SIZE);
+            float tu = static_cast<float>(m_tiles[i + j * width]->getTextureId() % (m_tileset->getSize().x / Constants::GRID_SIZE));
+            float tv = static_cast<float>(m_tiles[i + j * width]->getTextureId() / (m_tileset->getSize().x / Constants::GRID_SIZE));
 
             sf::Vertex* triangle = &m_vertices[(i + j * width) * 6];
 
-            triangle[0].position = sf::Vector2f(i * Constants::GRID_SIZE, j * Constants::GRID_SIZE);
-            triangle[1].position = sf::Vector2f((i + 1) * Constants::GRID_SIZE, j * Constants::GRID_SIZE);
-            triangle[2].position = sf::Vector2f((i + 1) * Constants::GRID_SIZE, (j + 1) * Constants::GRID_SIZE);
-            triangle[3].position = sf::Vector2f(i * Constants::GRID_SIZE, j * Constants::GRID_SIZE);
-            triangle[4].position = sf::Vector2f(i * Constants::GRID_SIZE, (j + 1) * Constants::GRID_SIZE);
-            triangle[5].position = sf::Vector2f((i + 1) * Constants::GRID_SIZE, (j + 1) * Constants::GRID_SIZE);
+            triangle[0].position = sf::Vector2f((float)(i * Constants::GRID_SIZE), (float)(j * Constants::GRID_SIZE));
+            triangle[1].position = sf::Vector2f((float)((i + 1) * Constants::GRID_SIZE), (float)(j * Constants::GRID_SIZE));
+            triangle[2].position = sf::Vector2f((float)((i + 1) * Constants::GRID_SIZE), (float)((j + 1) * Constants::GRID_SIZE));
+            triangle[3].position = sf::Vector2f((float)(i * Constants::GRID_SIZE), (float)(j * Constants::GRID_SIZE));
+            triangle[4].position = sf::Vector2f((float)(i * Constants::GRID_SIZE), (float)((j + 1) * Constants::GRID_SIZE));
+            triangle[5].position = sf::Vector2f((float)((i + 1) * Constants::GRID_SIZE), (float)((j + 1) * Constants::GRID_SIZE));
 
             triangle[0].texCoords = sf::Vector2f(tu * Constants::GRID_SIZE, tv * Constants::GRID_SIZE);
             triangle[1].texCoords = sf::Vector2f((tu + 1) * Constants::GRID_SIZE, tv * Constants::GRID_SIZE);
